Fixing the ExUnit async run logic
Muzak was not working on Elixir 1.12, I debuged it and it was basically
not requiring any tests, so there was no fail at all.

I couldn't find the exact reason for it, but I saw there was the introduction
of the function `ExUnit.async_run`, so I changed the `Task.async` to
call it and everything worked!